requirements.txt
bcrypt
dotenv
fastapi
logging
motor
pydantic
pyjwt
pymongo
uvicorn